T#784: Extract 10 dashboard/knowledge-overview routes to src/dashboard/routes.ts
Phase 2.6 of Library #102. Mechanical move, no logic changes.

Moved 10 routes: /api/dashboard, /summary, /activity, /growth +
/api/{stats,oracles,map,map3d,list,graph,reflect} knowledge-overview routes.

Helpers via DI: handleDashboardX, handleStats, handleReflect, handleList,
handleGraph, handleMap, handleMap3d, handleVectorStats, getSetting,
DB_PATH. oracleCache passed as state.

Server.ts: -100 lines. Cumulative T#771-T#784: -6,893 lines / -58%.
